refactor(site): driver dialogs as intercepting routes; fix RSC query boundaries
Phase C1b + a cross-cutting build fix.

C1b — driver CircuitDialog and SeasonDialog become URL-addressable modals via
Next.js parallel + intercepting routes:
- app/drivers/[driverRef]/layout.tsx adds a @dialog parallel slot;
  @dialog/(.)circuits/[circuitRef] and @dialog/(.)seasons/[year] intercept
  soft navigation into a RouteModal; non-intercepted circuits/[circuitRef] and
  seasons/[year] pages handle hard load / refresh / shared links
- CircuitDialogBody extracted (presentational, server-fed); Circuits/Career row
  links navigate to the dialog routes instead of opening a client-state dialog
- cached-data getDriverCircuitDialog feeds the dialog server-side; the dead
  CircuitDialog/SeasonDialog/useCircuitDialogData hook removed

RSC boundary fix (the build was red — cached-data is a server module imported by
layout, but it imported gql docs from client component files, dragging useState/
useRouter/useQuery into the server bundle):
- every cached-data-imported gql doc is now client-free: dead use*Data hooks
  reduced to their gql doc + types in place; .tsx component docs moved to
  co-located queries.ts; cached-data repointed to the client-free modules
- helpers.tsx no longer pulls next/router (dead useSlugs + dead useSeason removed)
- mark useChartSwitcher 'use client'

`next build` now compiles successfully (RSC boundaries + the parallel/intercepting
routes are valid); full prerender requires NEXT_PUBLIC_GRAPHQL_API_URL + a
reachable API, so runtime verification is pending an env with the API.
