+++ packages/site/src/app/drivers/[driverRef]/@dialog/default.tsx
@@ -1,0 +1,4 @@
+// No modal rendered for the base /drivers/[driverRef] route.
+export default function Default() {
+	return null;
+}
