+++ packages/site/src/components/page/constructor/season/Season.tsx
@@ -1,5 +1,4 @@
 import type { PropsWithChildren } from 'react';
-import { gql } from '@apollo/client';
 import { Alert, Grid, Link, Skeleton, Typography, type TypographyProps } from '@mui/material';
 import { DataGrid } from '@mui/x-data-grid';
 
@@ -7,6 +6,8 @@ import type { ConstructorSeasonRace } from '@/app/lib/cached-data';
 import { DriverByLine } from '@/components/app';
 import type { Team } from '@/gql/graphql';
 import { toPoints } from '@/helpers';
+
+export { ConstructorSeasonQuery } from './queries';
 
 const CellValueWrapper = ({
 	align = 'center',
@@ -16,31 +17,6 @@ const CellValueWrapper = ({
 		{children}
 	</Typography>
 );
-
-// Exported for cached-data.ts import
-export const ConstructorSeasonQuery = gql`
-	query ConstructorSeasonQuery($teamId: String!, $season: Int!) {
-		races(condition: { year: $season }, orderBy: ROUND_ASC) {
-			rowId
-			year
-			round
-			officialName
-			date
-			time
-
-			raceResults(condition: { teamId: $teamId }) {
-				raceId
-				gridPositionNumber
-				positionDisplayOrder
-				points
-				timeMillis
-				driverId
-				teamId
-				reasonRetired
-			}
-		}
-	}
-`;
 
 type SeasonProps = { teamId: Team['id']; season: number; races: ConstructorSeasonRace[] };
