+++ packages/site/src/components/page/season/index.ts
@@ -2,6 +2,4 @@ export { default as Schedule } from './Schedule';
 export { default as SeasonsList } from './SeasonsList';
 export { default as SeasonsQuery } from './SeasonsQuery';
 export type * from './types';
-export { default as useScheduleData } from './useScheduleData';
-export { default as useSeason } from './useSeason';
 export { default as useSeasonsList } from './useSeasonsList';
