+++ packages/site/src/components/page/season/standings/constructors/useConstructorsStandingsData.ts
@@ -1,10 +1,17 @@
 import { useCallback } from 'react';
-import { gql } from '@apollo/client';
 import { useSuspenseQuery } from '@apollo/client/react';
 
 import { useFallbackColor } from '@/components/ui';
 
 import type { Entity, RaceStandingsWithEntities } from '../charts';
+import type { SeasonTeamStandingNode } from './queries';
+import { constructorStandingsQuery } from './queries';
+
+export type {
+	SeasonTeamStandingColors,
+	SeasonTeamStandingNode,
+	SeasonTeamStandingTeamNode
+} from './queries';
 
 type ConstructorColors = {
 	primaryHex: string | null;
@@ -28,27 +35,6 @@ type RaceNode = {
 	raceTeamStandings: RaceConstructorStandingNode[];
 };
 
-// Extended colors for server-side display seed.
-export type SeasonTeamStandingColors = {
-	teamId: string;
-	primaryHex: string | null;
-	secondaryHex: string | null;
-};
-
-export type SeasonTeamStandingTeamNode = {
-	id: string;
-	name: string | null;
-	bio: { thumbnailUrl: string | null } | null;
-	colors: SeasonTeamStandingColors | null;
-};
-
-export type SeasonTeamStandingNode = {
-	teamId: string;
-	positionNumber: number | null;
-	points: string;
-	team: SeasonTeamStandingTeamNode | null;
-};
-
 type ConstructorStandingsQueryData = {
 	season: {
 		seasonTeamStandingsByYear: SeasonTeamStandingNode[];
@@ -68,48 +54,6 @@ const useMapConstructorToEntity = () => {
 		[fallbackColor]
 	);
 };
-
-export const constructorStandingsQuery = gql`
-	query constructorStandingsQuery($season: Int!) {
-		season(year: $season) {
-			year
-			seasonTeamStandingsByYear(orderBy: POSITION_NUMBER_ASC) {
-				teamId
-				positionNumber
-				points
-				team {
-					id
-					name
-					bio { thumbnailUrl }
-					colors {
-						teamId
-						primaryHex
-						secondaryHex
-					}
-				}
-			}
-			racesByYear(orderBy: ROUND_ASC) {
-				year
-				round
-				raceTeamStandings(orderBy: POSITION_NUMBER_ASC) {
-					raceId
-					teamId
-					engineManufacturerId
-					positionNumber
-					points
-					team {
-						id
-						name
-						colors {
-							teamId
-							primaryHex
-						}
-					}
-				}
-			}
-		}
-	}
-`;
 
 export default function useConstructorStandingsData(season: number) {
 	const { data } = useSuspenseQuery<ConstructorStandingsQueryData>(constructorStandingsQuery, {
