+++ packages/site/src/components/page/season/standings/drivers/useDriversStandingsData.ts
@@ -1,10 +1,13 @@
 import { useCallback } from 'react';
-import { gql } from '@apollo/client';
 import { useSuspenseQuery } from '@apollo/client/react';
 
 import { useFallbackColor } from '@/components/ui';
 
 import type { Entity, RaceStandingsWithEntities, StandingWithEntity } from '../charts';
+import type { SeasonDriverStandingNode } from './queries';
+import { driverStandingsQuery } from './queries';
+
+export type { SeasonDriverStandingDriverNode, SeasonDriverStandingNode } from './queries';
 
 type DriverColors = {
 	primaryHex: string | null;
@@ -36,33 +39,6 @@ type RaceNode = {
 	raceDriverStandings: RaceDriverStandingNode[];
 };
 
-// Extended fields fetched for server-side display seed (EntityDisplayProvider).
-type SeasonDriverStandingDriverColors = {
-	teamId: string;
-	primaryHex: string | null;
-	secondaryHex: string | null;
-};
-
-type SeasonDriverStandingEntrant = {
-	team: { id: string; colors: SeasonDriverStandingDriverColors | null } | null;
-};
-
-export type SeasonDriverStandingDriverNode = {
-	id: string;
-	firstName: string | null;
-	lastName: string | null;
-	abbreviation: string | null;
-	bio: { thumbnailUrl: string | null } | null;
-	seasonEntrantDrivers: SeasonDriverStandingEntrant[];
-};
-
-export type SeasonDriverStandingNode = {
-	driverId: string;
-	positionNumber: number | null;
-	points: string;
-	driver: SeasonDriverStandingDriverNode | null;
-};
-
 type DriverStandingsQueryData = {
 	season: {
 		seasonDriverStandingsByYear: SeasonDriverStandingNode[];
@@ -87,63 +63,6 @@ const useMapDriverToEntity = () => {
 		[fallbackColor]
 	);
 };
-
-export const driverStandingsQuery = gql`
-	query driverStandingsQuery($season: Int!) {
-		season(year: $season) {
-			year
-			seasonDriverStandingsByYear(orderBy: POSITION_NUMBER_ASC) {
-				year
-				driverId
-				positionNumber
-				points
-				driver {
-					id
-					firstName
-					lastName
-					abbreviation
-					bio { thumbnailUrl }
-					seasonEntrantDrivers(condition: {year: $season}, first: 1) {
-						team {
-							id
-							colors {
-								teamId
-								primaryHex
-								secondaryHex
-							}
-						}
-					}
-				}
-			}
-			racesByYear(orderBy: ROUND_ASC) {
-				year
-				round
-				raceDriverStandings(orderBy: POSITION_NUMBER_ASC) {
-					raceId
-					driverId
-					positionNumber
-					points
-					driver {
-						id
-						lastName
-						seasonEntrantDrivers(condition: {year: $season}, first: 1) {
-							year
-							driverId
-							teamId
-							team {
-								id
-								colors {
-									teamId
-									primaryHex
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-`;
 
 export default function useDriverStandingsData(season: number) {
 	const { data } = useSuspenseQuery<DriverStandingsQueryData>(driverStandingsQuery, {
